requirements.txt
peewee==2.8.5
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Login==0.3.2
Flask-WTF==0.13.1
Pillow==3.4.1
Pygments==2.1.3
pycrypto==2.6.1
